ordier的state同步为int，oomall_error_payment更名为oomall_error_account

+++ order/src/main/java/cn/edu/xmu/oomall/ooad201/order/model/po/OrderPo.java
@@ -172,7 +172,7 @@ public class OrderPo {
      *
      * @mbg.generated
      */
-    private Short state;
+    private Integer state;
 
     /**
      *
@@ -677,7 +677,7 @@ public class OrderPo {
      *
      * @mbg.generated
      */
-    public Short getState() {
+    public Integer getState() {
         return state;
     }
 
@@ -689,7 +689,7 @@ public class OrderPo {
      *
      * @mbg.generated
      */
-    public void setState(Short state) {
+    public void setState(Integer state) {
         this.state = state;
     }
